feat: split bff.ts into nested oclif directory (bff/{init,build,dev,validate}.ts)
- Extract shared types + helpers to bff/_shared.ts (oclif-skipped)
- bff/validate.ts: BffValidate + bffValidateCommand
- bff/build.ts: BffBuild + bffBuildCommand
- bff/dev.ts: BffDev + bffDevCommand
- bff/init.ts: BffInit + bffInitCommand (args: [name], flags: port, specs, no-static, project-version)
- Reduce bff.ts to re-export shim; all 65 existing BFF tests pass unmodified

Refs #94
